Break up big function that got model and instantiated
